Move GoogleOauth2 to different folder
